fix(frontend/ui): fix update banner false-positive, plugin list parsing, replace UnifiedLoadingView (#6775, #6774, #6698)
- cacheManagement.ts (#6775): guard buildHash must be a real string before
  comparing/storing. Backend SystemInfo does not emit buildHash, so it arrives
  as undefined; storing that as "undefined" caused the localStorage key to
  always differ, triggering the update banner on every page load.

- usePlugins.ts (#6774): handle both bare-array and {plugins,total} shapes.
  Backend PluginListResponse wraps the list; direct array access returned
  undefined, leaving plugins.value empty despite a successful API call.

- LoadingBoundary.vue + LoadingOverlay.vue (#6698): replace UnifiedLoadingView
  with two focused components. LoadingBoundary: spinner-or-slot (full replace)
  with #loading-message and #error-content named slots, timeout watcher, and
  loading-complete/loading-error/loading-timeout emits. LoadingOverlay: subtle
  pulsing badge on existing content while data refreshes. Both include
  Storybook stories documenting their contracts.

- Migrate all callers: App.vue, ChatInterface.vue, DesktopInterface.vue,
  PopoutChromiumBrowser.vue. Remove dead event handlers from ChatInterface
  (pure console.log stubs) and dead @loading-complete bindings from
  PopoutChromiumBrowser (static loading=true components unmount before the
  event could fire). Delete UnifiedLoadingView.vue and its stories.
